# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  errors.ApiError:
    properties:
      details: {}
      message:
        type: string
      status:
        type: integer
      type:
        $ref: '#/definitions/errors.ErrorType'
    type: object
  errors.ErrorType:
    enum:
    - VALIDATION_ERROR
    - NOT_FOUND_ERROR
    - UNAUTHORIZED_ERROR
    - FORBIDDEN_ERROR
    - CONFLICT_ERROR
    - INTERNAL_ERROR
    - BAD_REQUEST_ERROR
    - RATE_LIMIT_ERROR
    type: string
    x-enum-varnames:
    - ValidationError
    - NotFoundError
    - UnauthorizedError
    - ForbiddenError
    - ConflictError
    - InternalError
    - BadRequestError
    - RateLimitError
host: localhost:8080
info:
  contact:
    name: API Support
    url: http://www.swagger.io/support
  description: A high-performance license management system built with Go
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: License Management API
  version: "1.0"
paths:
  /audit-logs:
    get:
      consumes:
      - application/json
      description: Retrieve audit logs (Admins see all, users see only their own)
      parameters:
      - description: Page index (default 1)
        in: query
        name: pageIndex
        type: integer
      - description: Page size (default 20, max 100)
        in: query
        name: pageSize
        type: integer
      - description: Filter by action
        in: query
        name: action
        type: string
      - description: Filter by entity type
        in: query
        name: entity_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of audit logs
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/errors.ApiError'
      security:
      - BearerAuth: []
      summary: Get audit logs
      tags:
      - Audit
  /audit-logs/stats:
    get:
      consumes:
      - application/json
      description: Get aggregated statistics about audit logs (Admin only)
      produces:
      - application/json
      responses:
        "200":
          description: Audit log statistics
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/errors.ApiError'
      security:
      - BearerAuth: []
      summary: Get audit log statistics
      tags:
      - Audit
  /audit-logs/user:
    get:
      consumes:
      - application/json
      description: Retrieve audit logs specific to a user (Admin only)
      parameters:
      - description: User ID to filter logs
        in: query
        name: user_id
        required: true
        type: integer
      - description: Page index (default 1)
        in: query
        name: pageIndex
        type: integer
      - description: Page size (default 20, max 100)
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of user audit logs
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/errors.ApiError'
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/errors.ApiError'
      security:
      - BearerAuth: []
      summary: Get audit logs for a user
      tags:
      - Audit
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: JWT token in the format "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"